% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesm_bootstrap.R
\name{bayesm_bootstrap}
\alias{bayesm_bootstrap}
\title{Bayesian Marginal Structural Model Bootstrap Estimation}
\usage{
bayesm_bootstrap(
  wloglik_function,
  testdata,
  wmean = rep(1, 1000),
  nboot = 1000,
  inits1 = c(rep(0.1, length(treatment_vars) + 1), 4),
  optim_control = list(fnscale = -1),
  optim_method = "BFGS",
  optim_hessian = FALSE
)
}
\arguments{
\item{wloglik_function}{The weighted log-likelihood function to optimize.}

\item{testdata}{A data frame containing the outcome variable 'y' and treatment variables 'a_1', 'a_2', etc.}

\item{wmean}{A vector of weights to be used in the bootstrap procedure.}

\item{nboot}{The number of bootstrap iterations (default is 1000).}

\item{inits1}{Initial values for the optimization algorithm (default is c(rep(0.1, length(treatment_vars)+1),4).}

\item{optim_control}{List of control parameters to pass to the optim() function (default is list(fnscale = -1)).}

\item{optim_method}{The optimization method to be used in the optim() function (default is 'BFGS').}

\item{optim_hessian}{Logical; argument of optim() function (default is FALSE).}
}
\value{
A list of summary measures (mean, sd, quantile)
}
\description{
This function performs Bayesian non-parametric bootstrap to calculate causal effects
for Bayesian marginal structural models.
}
