% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesweight.R
\name{bayesweight}
\alias{bayesweight}
\title{Bayesian Treatment Effect Weight Estimation}
\usage{
bayesweight(
  trtmodel.list = list(a_1 ~ w1 + w2 + L1_1 + L2_1, a_2 ~ w1 + w2 + L1_1 + L2_1 + L1_2 +
    L2_2 + a_1),
  data = causaldata,
  n.chains = 1,
  n.iter = 25000,
  n.burnin = 15000,
  n.thin = 5,
  seed = 890123,
  parallel = FALSE
)
}
\arguments{
\item{trtmodel.list}{a list of formulas corresponding to each time point with the time-specific treatment variable on the left hand side and pre-treatment covariates to be balanced on the right hand side. The formulas must be in temporal order, and must contain all covariates to be balanced at that time point (i.e., treatments and covariates featured in early formulas should appear in later ones). Interactions and functions of covariates are allowed.}

\item{parallel}{}
}
\description{
Bayesian Treatment Effect Weight Estimation
}
\examples{

# Continuous outcome
testdata <- read.csv(system.file("extdata", "continuous_outcome_data.csv", package = "bayesmsm"))
weights <- bayesweight(trtmodel.list = list(a_1 ~ w1 + w2 + L1_1 + L2_1,
                                            a_2 ~ w1 + w2 + L1_1 + L2_1 + L1_2 + L2_2 + a_1),
                       data = testdata,
                       n.iter = 25000,
                       n.burnin = 15000,
                       n.thin = 5,
                       n.chains = 1,
                       seed = 890123,
                       parallel = FALSE)



}
