% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_effect.R
\name{calculate_effect}
\alias{calculate_effect}
\title{A function to calculate the effect of an intervention given the parameter estimates and intervention levels}
\usage{
calculate_effect(intervention_levels, variables, param_estimates)
}
\arguments{
\item{intervention_levels}{A numeric vector indicating the levels of intervention for each predictor variable.}

\item{variables}{A list of the names of the response variable and predictor variables extracted from the model.}

\item{param_estimates}{A vector of parameter estimates from the model.}
}
\value{
A numeric value representing the calculated effect of the specified intervention.
}
\description{
A function to calculate the effect of an intervention given the parameter estimates and intervention levels
}
