% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est_box.R
\name{plot_est_box}
\alias{plot_est_box}
\title{Error bar plots for causal treatment effects}
\usage{
plot_est_box(input, ...)
}
\arguments{
\item{input}{A data frame or model object containing bootstrap results.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A ggplot object presenting error bar plot of the mean effects and their 95\% credible intervals for comparator level, reference level, and ATE.
}
\description{
This function plots the point estimates and 95\% credible intervals of ATE and APO from bayesmsm output.
}
\examples{
testdata <- read.csv(system.file("extdata",
                     "sim_causal.csv",
                     package = "bayesmsm"))
model <- bayesmsm(ymodel = Y ~ A1 + A2 + A3,
                  nvisit = 3,
                  reference = c(rep(0,3)),
                  comparator = c(rep(1,3)),
                  treatment_effect_type = "sq",
                  family = "binomial",
                  data = testdata,
                  wmean = rep(1,500),
                  nboot = 10,
                  optim_method = "BFGS",
                  seed = 890123,
                  parallel = FALSE)
plot_est_box(model)
}
