% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesmsm.R
\name{summary.bayesmsm}
\alias{summary.bayesmsm}
\title{Title}
\usage{
\method{summary}{bayesmsm}(model)
}
\arguments{
\item{model}{A model object from bayesmsm}
}
\value{
A summary table of the results from bayesmsm.
}
\description{
Title
}
\examples{
# Continuous outcome
testdata <- read.csv(system.file("extdata", "continuous_outcome_data.csv", package = "bayesmsm"))
model <- bayesmsm(ymodel = y ~ a_1+a_2,
                           nvisit = 2,
                           reference = c(rep(0,2)),
                           comparator = c(rep(1,2)),
                           family = "gaussian",
                           data = testdata,
                           wmean = rep(1, 1000),
                           nboot = 1000,
                           optim_method = "BFGS",
                           seed = 890123,
                           parallel = TRUE,
                           ncore = 2)
summary.bayesmsm(model)



}
