% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bayesmsm.R
\name{summary_bayesmsm}
\alias{summary_bayesmsm}
\title{Summary function to generate result table from bayesmsm}
\usage{
summary_bayesmsm(model)
}
\arguments{
\item{model}{A model object from bayesmsm}
}
\value{
A summary table of the results from bayesmsm.
}
\description{
This function generates a ready to use result table that contents the estimated APO and ATE and their 95\% credible intervals
}
\examples{
testdata <- read.csv(system.file("extdata",
                     "sim_causal.csv",
                     package = "bayesmsm"))
model <- bayesmsm(ymodel = Y ~ A1 + A2 + A3,
                  nvisit = 3,
                  reference = c(rep(0,3)),
                  comparator = c(rep(1,3)),
                  treatment_effect_type = "sq",
                  family = "binomial",
                  data = testdata,
                  wmean = rep(1,500),
                  nboot = 10,
                  optim_method = "BFGS",
                  seed = 890123,
                  parallel = FALSE)
summary_bayesmsm(model)
}
