% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wloglik_normal.R
\name{wloglik_normal}
\alias{wloglik_normal}
\title{Weighted Log-Likelihood for Simple Linear Regression}
\usage{
wloglik_normal(param, Y, A, weight)
}
\arguments{
\item{param}{Numeric vector containing parameters (coefficients for treatment and standard deviation).}

\item{Y}{Numeric vector of dependent variable (outcome).}

\item{A}{Matrix of independent variables (treatment variables).}

\item{weight}{Numeric vector of treatment weights for each observation.}
}
\value{
Numeric value of the weighted log-likelihood.
}
\description{
This function calculates the weighted log-likelihood for a simple linear
regression model where the outcome y is continuous, and we assume the residual
follows a normal distribution.
}
