% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APO.R
\name{plot_APO}
\alias{plot_APO}
\title{Plot Average Potential Outcomes (APO)}
\usage{
plot_APO(input, effect_type, ...)
}
\arguments{
\item{input}{A data frame or model object containing bootstrap results.}

\item{effect_type}{A character string specifying which effect to plot (e.g., comparator or reference treatment sequences).}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A ggplot object representing density plot showing the distribution of the specified average potential outcome (reference or comparison).
}
\description{
This function plots the density of APO for a specified effect type from bayesmsm output.
}
\examples{
testdata <- read.csv(system.file("extdata",
                     "sim_causal.csv",
                     package = "bayesmsm"))
model <- bayesmsm(ymodel = Y ~ A1 + A2 + A3,
                  nvisit = 3,
                  reference = c(rep(0,3)),
                  comparator = c(rep(1,3)),
                  treatment_effect_type = "sq",
                  family = "binomial",
                  data = testdata,
                  wmean = rep(1,500),
                  nboot = 10,
                  optim_method = "BFGS",
                  seed = 890123,
                  parallel = FALSE)
plot_APO(model$bootdata, effect_type = "effect_comparator")
plot_APO(model, effect_type = "effect_reference")
}
