% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APO.R
\name{plot_APO}
\alias{plot_APO}
\title{Plot Average Potential Outcomes (APO)}
\usage{
plot_APO(input, effect_type, ...)
}
\arguments{
\item{input}{A data frame or model object containing 'bootdata', which include 'effect_comparator' and 'effect_reference' columns.}

\item{effect_type}{A character string specifying which effect to plot: 'effect_comparator' or 'effect_reference'.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A density plot showing the distribution of the specified average potential outcome (reference or comparison).
}
\description{
This function plots the density of APO for a specified effect type from bayesmsms output.
}
\examples{
testdata <- read.csv(system.file("extdata",
                                 "continuous_outcome_data.csv",
                                 package = "bayesmsm"))
model <- bayesmsm(ymodel = y ~ a_1+a_2,
                           nvisit = 2,
                           reference = c(rep(0,2)),
                           comparator = c(rep(1,2)),
                           treatment_effect_type = "sq",
                           family = "gaussian",
                           data = testdata,
                           wmean = rep(1, 1000),
                           nboot = 100,
                           optim_method = "BFGS",
                           parallel = FALSE,
                           ncore = 2)
plot_APO(model$bootdata, effect_type = "effect_comparator")
plot_APO(model, effect_type = "effect_reference")
}
