% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ATE.R
\name{plot_ATE}
\alias{plot_ATE}
\title{Plot Average Treatment Effect Density from bayesmsm output}
\usage{
plot_ATE(
  input,
  ATE = "RD",
  col_density = "blue",
  fill_density = "lightblue",
  main = "Posterior Predictive Distribution of Average Treatment Effect",
  xlab = "ATE",
  ylab = "Posterior Predictive Distribution",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{input}{A model object, data frame or vector containing the bootstrap estimates of ATE.}

\item{ATE}{define causal estimand of interest from RD, OR, RR.}

\item{col_density}{Color for the density plot (default is "blue").}

\item{fill_density}{Fill color for the density plot (default is "lightblue").}

\item{main}{Title of the plot (default is "Density of ATE Estimates").}

\item{xlab}{X-axis label (default is "ATE").}

\item{ylab}{Y-axis label (default is "Density").}

\item{xlim}{Limits for the x-axis (default is NULL).}

\item{ylim}{Limits for the y-axis (default is NULL).}

\item{...}{Additional graphical parameters passed to the plot function.}
}
\value{
A ggplot object representing the density plot for the posterior predictive distribution of the Average Treatment Effect (ATE).
}
\description{
This function plots the density of ATE from bayesmsm output.
}
\examples{
testdata <- read.csv(system.file("extdata",
                     "sim_causal.csv",
                     package = "bayesmsm"))
model <- bayesmsm(ymodel = Y ~ A1 + A2 + A3,
                  nvisit = 3,
                  reference = c(rep(0,3)),
                  comparator = c(rep(1,3)),
                  treatment_effect_type = "sq",
                  family = "binomial",
                  data = testdata,
                  wmean = rep(1,500),
                  nboot = 10,
                  optim_method = "BFGS",
                  seed = 890123,
                  parallel = FALSE)
plot_ATE(model)
}
