% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ATE.R
\name{plot_ATE}
\alias{plot_ATE}
\title{Plot Average Treatment Effect (ATE) Density from Bootstrap Results}
\usage{
plot_ATE(
  model,
  estimand = c("RD", "RR", "OR"),
  col_density = "blue",
  fill_density = "lightblue",
  main = "Posterior Predictive Distribution of Average Treatment Effect (ATE)",
  xlab = "ATE",
  ylab = "Posterior Predictive Distribution",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{col_density}{Color for the density plot (default is "blue").}

\item{fill_density}{Fill color for the density plot (default is "lightblue").}

\item{main}{Title of the plot (default is "Density of ATE Estimates").}

\item{xlab}{X-axis label (default is "ATE").}

\item{ylab}{Y-axis label (default is "Density").}

\item{xlim}{Limits for the x-axis (default is NULL).}

\item{ylim}{Limits for the y-axis (default is NULL).}

\item{...}{Additional graphical parameters passed to the plot function.}

\item{input}{A model object, data frame or vector containing the bootstrap estimates of ATE.}
}
\description{
Plot Average Treatment Effect (ATE) Density from Bootstrap Results
}
\examples{
testdata <- read.csv(system.file("extdata", "continuous_outcome_data.csv", package = "bayesmsm"))
model <- bayesmsm(ymodel = y ~ a_1+a_2,
                           nvisit = 2,
                           reference = c(rep(0,2)),
                           comparator = c(rep(1,2)),
                           family = "gaussian",
                           data = testdata,
                           wmean = rep(1, 1000),
                           nboot = 1000,
                           optim_method = "BFGS",
                           parallel = TRUE,
                           ncore = 2)
plot_ATE(model)

}
