% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est_box.R
\name{plot_est_box}
\alias{plot_est_box}
\title{Error bar plots for treatment effects}
\usage{
plot_est_box(input, ...)
}
\arguments{
\item{input}{A data frame or model object containing 'bootdata', which include 'effect_comparator', 'effect_reference', and 'ATE' columns.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
An error bar plot of the mean effects and their 95\% confidence intervals for comparator level, reference level, and ATE.
}
\description{
Error bar plots for treatment effects
}
\examples{
plot_est_box(model$bootdata) # without reference & comparator information below labels
plot_est_box(model) # with reference & comparator information below labels

}
