% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bayesmsm.R
\name{summary_bayesmsm}
\alias{summary_bayesmsm}
\title{Summary function to generate result table from bayesmsm}
\usage{
summary_bayesmsm(model)
}
\arguments{
\item{model}{A model object from bayesmsm}
}
\value{
A summary table of the results from bayesmsm.
}
\description{
Summary function to generate result table from bayesmsm
}
\examples{
# Continuous outcome
testdata <- read.csv(system.file("extdata", "continuous_outcome_data.csv", package = "bayesmsm"))
model <- bayesmsm(ymodel = y ~ a_1+a_2,
                           nvisit = 2,
                           reference = c(rep(0,2)),
                           comparator = c(rep(1,2)),
                           treatment_effect_type = "sq",
                           family = "gaussian",
                           data = testdata,
                           wmean = rep(1, 1000),
                           nboot = 10,
                           optim_method = "BFGS",
                           seed = 890123,
                           parallel = TRUE,
                           ncore = 2)
summary_bayesmsm(model)
}
